部屋がないのに、入室判定の場合にバグになってしまうので、enter?で部屋がほんとにあるか確認

+++ app/helpers/entrances_helper.rb
@@ -25,13 +25,22 @@ module EntrancesHelper
     end
   end
   def enter?
-    !current_user.room_id.nil?
+    if !current_user.room_id.nil?
+      if current_room
+        true
+      else
+        exit
+        false
+      end
+    else
+      false
+    end
   end
   def exit
     exit_room = current_room
     current_user.room_id = nil
     current_user.save
-    if exit_room.users.count == 0
+    if !exit_room.nil? && exit_room.users.count == 0
       exit_room.destroy
     end
     @current_room = nil
